refactor(invalid_upcast_comparisons): move to under `operators`
and clean up a bit:
- put the lint functions in the order they're called
- simplify `upcast_comparison_bounds_err`
  - It first handled `Eq`/`Ne`, then one pair of cases with `Lt`/Le`,
    then another pair of cases with `Lt`/`Le` -- turn that into a single
    `match rel`.
  - It used a pattern of `if invert { a } else { b }`, which I found to
    be hard because of the need to keep in mind the implicit `invert &&`
    and `!invert &&` when looking at `a` and `b`. The form
    `(invert && a) || (!invert && b)` should be a bit more explicit,
    while compiling down to the same thing
  - The conditions were already partially written in the order `lb`,
    `norm_rhs_val`, `ub` -- change the ones that weren't
